Started to move TPortal.subs to Subs.php and Portal.php, broken currently
