+++ package-info.xml
@@ -19,7 +19,6 @@
 		<remove-dir name="THEMEDIR/images/tinyportal" />
 		<remove-dir name="BOARDDIR/TinyPortal" />
 		<remove-dir name="THEMEDIR/scripts/tinyportal" />
-		<remove-file name="SUBSDIR/TPortal.subs.php" />
 		<remove-file name="LANGUAGEDIR/english/TParticle.english.php" />
 		<remove-file name="LANGUAGEDIR/english/TPhelp.english.php" />
 		<remove-file name="LANGUAGEDIR/english/TPmodules.english.php" />
